fix overflow caused by casting

+++ core/src/main/java/org/bitcoinj/net/ConnectionHandler.java
@@ -49,7 +49,7 @@ public class ConnectionHandler implements MessageWriteTarget {
     private static final int BUFFER_SIZE_LOWER_BOUND = 4096;
     private static final int BUFFER_SIZE_UPPER_BOUND = 65536;
 
-    private static final int OUTBOUND_BUFFER_BYTE_COUNT = Message.MAX_SIZE + 24; // 24 byte message header
+    private static final long OUTBOUND_BUFFER_BYTE_COUNT = Message.MAX_SIZE + 24L; // 24 byte message header
 
     // We lock when touching local flags and when writing data, but NEVER when calling any methods which leave this
     // class into non-Java classes.
